+++ project/accounts/models.py
@@ -1,5 +1,8 @@
 from django.contrib.auth.models import AbstractUser
+from django.db import models
+from django.utils.translation import gettext_lazy as _
 
 
 class User(AbstractUser):
-    pass
+    email = models.EmailField(_("Email Address"), unique=True)
+    email_verified = models.BooleanField(default=False)