+++ project/accounts/views/__init__.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.conf import settings
+from django.contrib.auth import login
 from accounts.forms import RegisterForm
 
 
@@ -9,9 +10,10 @@ def register(request):
         if request.method == "POST":
             form = RegisterForm(request.POST)
             if form.is_valid():
-                form.save()
-                return redirect("login")
+                user = form.save()
+                login(request, user)
+                return redirect("accounts:verify_email_send_token")
         context = {"form": form}
     else:
         context = {"not_allowed": True}
-    return render(request, "register.html", context)
+    return render(request, "registration/register.html", context)